requirements.txt
Pillow
markdown
requests
beautifulsoup4
google-api-python-client==1.7.2
PyPDF2
instructor
fastapi
pydantic
python-dotenv
uvicorn
google-auth==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1